requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
python-telegram-bot>=20.0
loguru>=0.7.0
scikit-learn>=1.3.0
